Revert "create base textarea, create cv, upd"
