+++ components/auth/LoginForm.vue
@@ -73,7 +73,7 @@
 	import { useRouter } from '#app';
 	import { ref } from 'vue';
 	import { useAuthValidation } from '~/composables/useAuthValidation';
-	import { login } from '~/services/auth';
+	import { login } from '~/utils/graphql-services/auth-service';
 
 	const router = useRouter();
 	const isPasswordVisible = ref(false);
