+++ components/auth/SignUpForm.vue
@@ -87,7 +87,7 @@
 	import { useRouter } from '#app';
 	import { ref } from 'vue';
 	import { useAuthValidation } from '~/composables/useAuthValidation';
-	import { signUp } from '~/services/auth';
+	import { signUp } from '~/utils/graphql-services/auth-service';
 
 	const router = useRouter();
 	const isPasswordVisible = ref(false);
