--- graphql/mutations/cv.graphql
@@ -1,8 +1,0 @@
-mutation CreateCV($cv: CreateCvInput!) {
-    createCv(cv: $cv) {
-        id
-        name
-        education
-        description
-    }
-}
