+++ pages/cvs/index.vue
@@ -1,29 +1,7 @@
 <template>
 	<div class="h-[calc(100vh-0.75rem-90px)]">
-		<div class="flex items-center justify-between">
-			<div>
-				<BaseSearchBar
-					v-model="searchQuery"
-					placeholder="Search"
-					class="mb-5"
-				/>
-			</div>
-			<div>
-				<BaseButton
-					color="primary"
-					type="button"
-					variant="text"
-					@click="handleAddCV"
-				>
-					<div class="flex items-center justify-center gap-3">
-						<PlusIcon
-							color="var(--color-button-primary-text-text)"
-							width="24"
-						/>
-						CREATE CV
-					</div>
-				</BaseButton>
-			</div>
+		<div>
+			<BaseSearchBar v-model="searchQuery" placeholder="Search" class="mb-5" />
 		</div>
 		<div>
 			<Table
@@ -33,36 +11,11 @@
 				row-component="TableCvRow"
 			/>
 		</div>
-		<BaseModal
-			v-model:is-open="isAddCVModalOpen"
-			confirm-text="CREATE"
-			cancel-text="CANCEL"
-			title="Create CV"
-			:has-changes="!!name"
-			@confirm="handleAddCVConfirm"
-		>
-			<BaseInput id="name" v-model="name" type="text" label="Name" />
-			<BaseInput
-				id="education"
-				v-model="education"
-				type="text"
-				label="Education"
-			/>
-			<BaseTextarea
-				id="description"
-				v-model="description"
-				label="Description"
-				:rows="6"
-			/>
-		</BaseModal>
 	</div>
 </template>
 
 <script setup lang="ts">
-	import { useRouter } from '#app';
-	import PlusIcon from '~/components/icons/PlusIcon.vue';
-	import { createCV, getAllCvs } from '~/services/cv';
-	import { showErrorToast, showSuccessToast } from '~/utils/toast/toast';
+	import { getAllCvs } from '~/utils/graphql-services/cv-service';
 
 	interface Cv {
 		id: string | number;
@@ -75,25 +28,9 @@
 		};
 	}
 
-	interface CreateCV {
-		userId: string;
-		name: string;
-		education: string;
-		description: string;
-	}
-
 	const searchQuery = ref('');
 	const isDataLoaded = ref(false);
 	const cvs = ref<Cv[]>([]);
-	const isAddCVModalOpen = ref(false);
-	const name = ref('');
-	const education = ref('');
-	const description = ref('');
-	const router = useRouter();
-
-	const userId = computed(() => {
-		return cvs.value.length > 0 ? String(cvs.value[0].user.id) : '';
-	});
 
 	const headers = reactive([
 		{ key: 'name', label: 'CV Name', isSortable: true },
@@ -135,34 +72,6 @@
 			console.error('Error loading CVs:', error);
 		} finally {
 			isDataLoaded.value = true;
-		}
-	};
-
-	const handleAddCV = () => {
-		name.value = '';
-		education.value = '';
-		description.value = '';
-		isAddCVModalOpen.value = true;
-	};
-
-	const handleAddCVConfirm = async () => {
-		if (!name.value) return;
-		try {
-			const cv: CreateCV = {
-				userId: userId.value,
-				name: name.value,
-				education: education.value,
-				description: description.value,
-			};
-			const data = await createCV(cv);
-			console.log(data);
-			if (data) {
-				isAddCVModalOpen.value = false;
-				router.push(`/cvs/${data.data.createCv.id}/details`);
-				showSuccessToast('CV create successfully');
-			}
-		} catch (error) {
-			showErrorToast('CV create failed');
 		}
 	};
